[TASK] Moving StandaloneAssetPathViewHelper (Highlighter VH namespace)
