+++ Resources/Private/Partials/PageAssets/SyntaxHighlighter.html
@@ -52,7 +52,7 @@
 	<f:section name="Syntaxhighlighter">
 		SyntaxHighlighter.autoloader(
 		<f:for each="{brushes}" key="brushAlias" as="brushIdentifier" iteration="i">
-			{f:if(condition: '{i.index} > 0', then: ',')}'{brushAlias} {boc:standaloneAssetPath(baseUrl: '{boc:makeAbsolute(url: settings.baseUrl)}', resourcePath: settings.scripts, type: 'scripts')}shBrush{brushIdentifier}.js'
+			{f:if(condition: '{i.index} > 0', then: ',')}'{brushAlias} {boc:highlighter.syntaxHighlighter.assetPath(baseUrl: '{boc:makeAbsolute(url: settings.baseUrl)}', resourcePath: settings.scripts, type: 'scripts')}shBrush{brushIdentifier}.js'
 		</f:for>
 		);
